Limit user by credit used on AWS
Download price from aws endpoint.
Add a User service function for search history of a user with a specified driver.
Add a machineType key on history model.
Add a function for calculate the user credit
Add function how calculate the credit for a user
getUserCredit use getUserHistory for know the machines used
It search for price of all machine and return a global price
Add a User service function for search history of a user with a specified driver.
Add a machineType key on history model.
Add test for aws getUserCredit function
Test if the function return a valid price, using a fake price data
Add credit to user model
Use the sessionEnded function to check the credit used byt the user
and update it in the database
Initialise price with the current Aws region
Add the front end part of limt credit
Add a 402 error code when credit is exceeded