Limit user by credit used on AWS
Download price from aws endpoint.
Add a User service function for search history of a user with a specified driver.
Add a machineType key on history model.
Add a function for calculate the user credit
Add function how calculate the credit for a user
getUserCredit use getUserHistory for know the machines used
It search for price of all machine and return a global price
Add a User service function for search history of a user with a specified driver.
Add a machineType key on history model.
Add test for aws getUserCredit function
Test if the function return a valid price, using a fake price data
Add credit to user model
Use the sessionEnded function to check the credit used byt the user
and update it in the database
Initialise price with the current Aws region
Add the front end part of limt credit
Add a 402 error code when credit is exceeded

Add the history model on backend.

Add unit test for machine service.

Change the iaas on the Config service test.

AWS price

Download price from aws endpoint.
Add a User service function for search history of a user with a specified driver.
Add a machineType key on history model.

Add getUserCredit function (not finished)

Add a function for calculate the user credit

Add function how calculate the credit for a user

getUserCredit use getUserHistory for know the machines used
It search for price of all machine and return a global price

Add machine-type, and flavor

add flavor for the machine size

Add some modification for more visibility,

using JSON.parse on the data of the fake file price for the driver test.
Use an array on machine controller, for corresponding to frontend.

Test the err variable when aws endpoint is called.
Supress duplicate key.

remove all console log

Add modification test for corresponding
to new aws api/apps/connections end point

Indent

Add more visibility

User function getUserHistory return the right time of duration of the user machines history.

Aws driver add a function getAwsPrice to parse the price receved on aws endpoint.
getUserCredit now don't need to parse or serach price, it's already do in getAwsPrice.

Test cover all possibility of histories entrie.

withing lodash

getUserHistory return the different with the current time if their is no endDate
It add 1 hour only if the history duration is not exactly equal to 1 hour

Add tests for this 2 thinks

Add comment for more visiblity

Add the front part for aws price config.
If not using aws, can't configure credit limit

Add front-end part for set credit used by a user

User how not use aws, can't see their credit.

The awsPrice is now a promise,
Front end work correctly with credit
remove comment

Disable the option for admin to set credit of a user

User credit is now only for the actual month (month start the day he sign-up)

Add test to check this.
Change start/end date on all test, to match with this too.

Price calculated can't have more than 4 decimales.

remove a log

remove a </tr>

Fix bug: if credit limit is set to 0, no vdi can be up.

Fix bug initialize taking to long

Add 5 min on tests time for getUserHistory, for be sur it take the right timer